# backend/openapi.yaml
openapi: 3.0.3
info:
  title: MedLine API
  description: >
    Middleware de gestión de farmacia hospitalaria: RBAC, inventario con FEFO
    y Stock Útil, libro de movimientos, Motor de Compra Inteligente, gestión
    de proveedores, devoluciones/logística inversa, historia clínica
    (farmacia) con versionado, y auditoría inmutable.


    Además de los endpoints REST documentados aquí, el servidor expone un
    canal **WebSocket** en `ws://<host>/ws?token=<JWT>` para notificaciones en
    tiempo real (nuevos movimientos, cambios de estado de órdenes de compra,
    devoluciones procesadas/rechazadas). OpenAPI no modela WebSockets; se deja
    esta nota como referencia. Cada evento llega sólo a los sockets cuyo
    usuario tiene el permiso del recurso correspondiente.
  version: "1.0.0"
servers:
  - url: /api
    description: API relativa (mismo origen que el frontend en desarrollo/producción)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token emitido por POST /auth/login. Embebe el set de permisos del
        usuario; el backend lo revalida en cada endpoint protegido con
        requirePermission(...).
  schemas:
...
    Movement:
      type: object
      properties:
        id: { type: string }
        timestamp: { type: string, format: date-time }
        quantityDelta: { type: integer, description: 'Positivo = entrada, negativo = salida' }
        movementType: { type: string, enum: [ENTRY, EXIT, ADJUSTMENT, RETURN, INTERNAL_TRANSFER] }
        sourceLocation: { type: string, nullable: true }
        destinationLocation: { type: string, nullable: true }
        note: { type: string, nullable: true }
        item:
          type: object
          properties: { id: { type: string }, name: { type: string } }
        user:
          type: object
          properties: { id: { type: string }, fullName: { type: string } }
...
  parameters:
    ItemIdFilter:
      name: itemId
      in: query
      schema: { type: string }
      description: Filtra por un ítem específico
    SearchQuery:
      name: q
      in: query
      schema: { type: string }
      description: Búsqueda de texto libre (multi-campo, contains case-insensitive en SQLite)
security:
  - bearerAuth: []
paths:
...
  /movements:
    get:
      tags: [Movements]
      summary: Libro de movimientos, con filtro por ítem y búsqueda de texto libre (requiere movements:read)
      parameters:
        - $ref: '#/components/parameters/ItemIdFilter'
        - $ref: '#/components/parameters/SearchQuery'
      responses:
        '200':
          description: Últimos 200 movimientos
          content:
            application/json:
              schema:
                type: object
                properties:
                  movements: { type: array, items: { $ref: '#/components/schemas/Movement' } }
    post:
      tags: [Movements]
      summary: Registrar un movimiento y ajustar el lote afectado (requiere movements:create)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemId, quantityDelta, movementType]
              properties:
                itemId: { type: string }
                batchId: { type: string }
                quantityDelta: { type: integer }
                movementType: { type: string, enum: [ENTRY, EXIT, ADJUSTMENT, RETURN, INTERNAL_TRANSFER] }
                sourceLocation: { type: string }
                destinationLocation: { type: string }
                note: { type: string }
      responses:
        '201': { description: Movimiento registrado }
        '409': { description: Stock insuficiente en el lote }
...